Add DeltaE mode
